backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
